List: fix uncorrect scrolling behaviour
Fixed an uncorrect scrolling behaviour of the list. It sometimes caused also a segmentation fault, accessing to its children array as out of index.

TODO: now it build and render 1 items more when the number of items to be displayed is even. It should have a little different logic than the odd version, since we should consider that in this case the interval is not perfectly centered.