New navigation system
There is a new navigation system that solved the issues of the previous one, which was affected by a circular dependency.

Now there is a ViewFactor that centralizes the creations of the views, and the navigator creates the view through it if needed. The user must implement the IViewFactory interface, defining for each view its unique ID.

In that way, a view doesn't need anymore to know (and instantiate) the other views it want to navigate to. It just need to pass to the navigator the ID of the view it want to navigate to. This approach enable a less coupled system. Moreover, a ViewFactory abstract the creation of the view, so it could be useful when integrating the framework with another design pattern (e.g. MVVM), since you can easily inject dependencies inside the view during its creation controlled by the developer.