[Feat] EventLoop + Periodic Tasks scheduling
This commit introduces the event loop to the ASUX framework, enabling a refresh rate mechanism for the rendering of the UI. The main loop is no longer blocked by the reading of inputs from terminal.

A scheduler for periodic tasks has been added too, along with the Runnable interface, allowing developers to define logic they want to executes at a specified frame rate.

NOTE: Maybe some work still need to be done to improve the accuracy of the tasks timing/scheduling.